S6_Streamlit_for_AI_Apps/S6_requirements.txt
streamlit
langchain-core
langchain-openai
langchain-google-genai
python-dotenv
